Even tinier data types

+++ kfmon.h
@@ -72,6 +72,7 @@
 #	define KFMON_CONFIGPATH "/home/niluje/Kindle/Staging/kfmon"
 #endif
 
+// NOTE: See https://kernelnewbies.org/FAQ/DoWhile0 for the reasoning behind the use of GCC's ({ … }) notation
 // Log everything to stderr (which actually points to our logfile)
 #define LOG(prio, fmt, ...)                                                                                            \
 	({                                                                                                             \
@@ -146,6 +147,7 @@ typedef struct
 } WatchConfig;
 
 // Hardcode the max amount of watches we handle
+// NOTE: Cannot exceed INT8_MAX!
 #define WATCH_MAX 16
 
 // Used to keep track of our spawned processes, by storing their pids, and their watch idx.
@@ -153,14 +155,14 @@ typedef struct
 // As well as issue #2 for details of past failures w/ a SIGCHLD handler
 struct process_table
 {
-	pid_t     spawn_pids[WATCH_MAX];
-	short int spawn_watchids[WATCH_MAX];
+	pid_t  spawn_pids[WATCH_MAX];
+	int8_t spawn_watchids[WATCH_MAX];
 } PT;
-pthread_mutex_t  ptlock = PTHREAD_MUTEX_INITIALIZER;
-static void      init_process_table(void);
-static short int get_next_available_pt_entry(void);
-static void      add_process_to_table(short int, pid_t, unsigned short int);
-static void      remove_process_from_table(short int);
+pthread_mutex_t ptlock = PTHREAD_MUTEX_INITIALIZER;
+static void     init_process_table(void);
+static int8_t   get_next_available_pt_entry(void);
+static void     add_process_to_table(int8_t, pid_t, uint8_t);
+static void     remove_process_from_table(int8_t);
 
 static void init_fbink_config(void);
 
@@ -197,7 +199,7 @@ static int  watch_handler(void*, const char*, const char*, const char*);
 static bool validate_watch_config(void*);
 static int  load_config(void);
 // Ugly global. Remember how many watches we set up...
-unsigned short int watch_count = 0;
+uint8_t watch_count = 0;
 // Make our config global, because I'm terrible at C.
 #pragma GCC diagnostic push
 #pragma GCC diagnostic ignored "-Wmissing-braces"
@@ -207,14 +209,14 @@ FBInkConfig            fbink_config            = { 0 };
 #pragma GCC diagnostic push
 
 static unsigned int qhash(const unsigned char*, size_t);
-static bool         is_target_processed(unsigned short int, bool);
+static bool         is_target_processed(uint8_t, bool);
 
 void*        reaper_thread(void*);
-static pid_t spawn(char* const*, unsigned short int);
+static pid_t spawn(char* const*, uint8_t);
 
-static bool  is_watch_already_spawned(unsigned short int);
+static bool  is_watch_already_spawned(uint8_t);
 static bool  is_blocker_running(void);
-static pid_t get_spawn_pid_for_watch(unsigned short int);
+static pid_t get_spawn_pid_for_watch(uint8_t);
 
 static bool handle_events(int);
